More shuffling functions and imports
